requirements.txt
Flask
sqlalchemy
psycopg2
pycryptodome
python-dotenv
Flask-SQLAlchemy
